labs/Laboratorio6/main.py
from dotenv import load_dotenv
from langchain_experimental.agents import create_csv_agent
from langchain_openai import ChatOpenAI
from langchain.agents import create_react_agent, AgentExecutor

# Esto lo que hace es generar un agente python para generar codigo automaticamente
from langchain_experimental.tools import PythonREPLTool

