labs/Laboratorio6/requirements.txt
langchain ==0.3.6
langchain-experimental ==0.3.2
langchainhub
langchain-community
langchain-text-splitters
langchain-openai
langsmith
langchain-core
langgraph
langchain-pinecone
requests
black
openai
python-dotenv
pinecone
beautifulsoup4
altair<5
firecrawl-py
qrcode
streamlit
pandas
tabulate